Simplify staking reward pipeline and make a new begin block helper (#951)
* Simplify staking reward pipeline and make a new begin block helper

* Move helper methods to single file

* Don't correct accumulation store at migration time
